fix(product): split submit-stamp repin out of use-transcript-stick-to-bottom (repo-shape)
use-transcript-stick-to-bottom.ts sat at 431 lines, over the 400-line
LARGE_FRONTEND_FILE soft cap with no allowlist entry. Extracts the
session-scoped submit-stamp re-pin effect (PRO-175) into its own colocated
hook, use-transcript-submit-stamp-repin.ts, taking the minimal
{lastPromptSubmittedAtMs, sessionKey, setPinned, scrollToBottom,
startGlueLoop} contract. useTranscriptStickToBottom's public API and
behavior are unchanged; the session-stamp test still exercises it only
through that public hook.
