+++ apps/packages/product-client/src/hooks/chat/ui/use-transcript-stick-to-bottom.ts
@@ -9,6 +9,7 @@ import {
   TRANSCRIPT_USER_SCROLL_SETTLE_MS,
   type TranscriptScrollSample,
 } from "#product/hooks/chat/ui/transcript-row-list-model";
+import { useTranscriptSubmitStampRepin } from "#product/hooks/chat/ui/use-transcript-submit-stamp-repin";
 import { useTranscriptUserScrollIntent } from "#product/hooks/chat/ui/use-transcript-user-scroll-intent";
 
 function interactionNow(): number {
@@ -301,51 +302,18 @@ export function useTranscriptStickToBottom({
     glueFrameRef.current = requestAnimationFrame(tick);
   }, [scrollRef, scrollToBottom]);
 
-  // A prompt submit is an explicit return-to-bottom intent: re-pin even when
-  // the pin was silently lost earlier (so the sent bubble can never render
-  // clipped behind the dock), snap, and glue across the composer-collapse /
-  // row-measurement settle so the multi-frame geometry change lands as one
-  // silent jump, exactly like session re-entry. Unlike the scroll-to-bottom
-  // button, a submit does NOT consume the manual-only overlay range: the
-  // follow target stays the soft bottom above any dock-slot card, so the
-  // stream never slides under it (a range the user already consumed stays
-  // consumed until they scroll away). Registered after the inset effect above but
-  // before consumer layout effects, so their pinned snaps read the restored
-  // pin. Only a monotonic increase of the submission stamp qualifies — see
-  // the option's contract.
-  //
-  // PRO-175: the row lists never remount on a session switch, so this ref
-  // would otherwise carry the PREVIOUS session's stamp across the switch. A
-  // revisited session's own (unrelated, possibly old) stamp then looks like a
-  // fresh increase and fires a spurious re-pin/snap/glue with zero new
-  // content. `sessionKey` scopes the comparison to session identity: a
-  // session-boundary crossing re-baselines the ref to the incoming session's
-  // CURRENT stamp (not null — nulling it would make the very next run see
-  // previous == null and misfire through the other door) and skips the
-  // compare for that render. `resetForSession` still unconditionally
-  // pins/snaps/glues on every switch by design; this only removes the SECOND,
-  // redundant re-pin the stale stamp used to trigger alongside it.
-  const lastPromptSubmittedAtRef = useRef(lastPromptSubmittedAtMs);
-  const sessionKeyRef = useRef(sessionKey);
-  useLayoutEffect(() => {
-    const previousSessionKey = sessionKeyRef.current;
-    sessionKeyRef.current = sessionKey;
-    const previous = lastPromptSubmittedAtRef.current;
-    lastPromptSubmittedAtRef.current = lastPromptSubmittedAtMs;
-    if (previousSessionKey !== undefined && previousSessionKey !== sessionKey) {
-      // Session boundary: the ref above is now the new baseline. Do not
-      // compare against the outgoing session's stamp.
-      return;
-    }
-    if (
-      lastPromptSubmittedAtMs != null
-      && (previous == null || lastPromptSubmittedAtMs > previous)
-    ) {
-      setPinned(true);
-      scrollToBottom();
-      startGlueLoop();
-    }
-  }, [lastPromptSubmittedAtMs, scrollToBottom, sessionKey, setPinned, startGlueLoop]);
+  // A prompt submit is an explicit return-to-bottom intent (PRO-175 scopes it
+  // to session identity so a session switch can't misfire it) — see
+  // use-transcript-submit-stamp-repin.ts. Registered after the inset effect
+  // above but before consumer layout effects, so their pinned snaps read the
+  // restored pin.
+  useTranscriptSubmitStampRepin({
+    lastPromptSubmittedAtMs,
+    sessionKey,
+    setPinned,
+    scrollToBottom,
+    startGlueLoop,
+  });
 
   // Session re-entry: snap instantly, then glue for a few frames so the
   // measurement backlog of freshly mounted rows (virtualizer estimates
